requirements.txt
addict==2.2.1
Beaker==1.11.0
bottle==0.12.18
cheroot==8.4.2
fiona==1.8.13
gdal==2.3.2
gdal2tiles==0.1.7
geopandas==0.8.1
joblib==0.16.0
pandas==1.0.5
pyyaml>4.2b1
rasterio==1.1.5
rpyc==4.1.5
Shapely==1.7.0
scikit-image==0.16.2
torch==1.6.0
torchvision==0.7.0
utm==0.5.0
tornado==5.1.0
pybind11
scikit-build>=0.11.1
cmake>=3.14.2
opencv-python>=4.4
